Sync: realtime logging for git sync
