+++ lib/util.js
@@ -446,7 +446,6 @@ const util = {
   },
 
   setGitVersion: async (gitRepoPath, version, alwaysReset = false) => {
-    await util.fetch(gitRepoPath)
     await util.runGitAsync(gitRepoPath, ['clean', '-f'])
     let shouldReset = alwaysReset
     if (!shouldReset) {
